Renaming raft runtime -> client api
