+++ conda/recipes/libraft_client_api/meta.yaml
@@ -8,7 +8,7 @@
 {% set cuda_major=cuda_version.split('.')[0] %}
 {% set ucx_py_version=environ.get('UCX_PY_VERSION') %}
 package:
-  name: libraft-runtime
+  name: libraft-client-api
   version: {{ version }}
 
 source:
@@ -58,4 +58,4 @@ about:
   home: http://rapids.ai/
   license: Apache-2.0
   # license_file: LICENSE
-  summary: libraft-runtime library
+  summary: libraft-client-api library
