+++ conda/recipes/libraft_distance/meta.yaml
@@ -38,7 +38,7 @@ requirements:
   build:
     - cmake>=3.20.1
   host:
-    - libraft-runtime {{ version }}
+    - libraft-client-api {{ version }}
     - nccl>=2.9.9
     - cudatoolkit {{ cuda_version }}.*
     - ucx-py {{ ucx_py_version }}
@@ -47,7 +47,7 @@ requirements:
     - gmock
     - librmm {{ minor_version }}
   run:
-    - libraft-runtime {{ version }}
+    - libraft-client-api {{ version }}
     - nccl>=2.9.9
     - ucx-py {{ ucx_py_version }}
     - ucx-proc=*=gpu
