+++ cpp/include/raft_client/error.hpp
