+++ cpp/include/raft_client/linalg/cublas_macros.hpp
@@ -20,7 +20,7 @@
 #pragma once
 
 #include <cublas_v2.h>
-#include <raft_runtime/error.hpp>
+#include <raft_client/error.hpp>
 
 ///@todo: enable this once we have logger enabled
 //#include <cuml/common/logger.hpp>
