+++ cpp/include/raft_client/raft.hpp
